% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfb_team_game_by_game.R
\name{.cfb_team_game_by_game}
\alias{.cfb_team_game_by_game}
\title{Extract Team Stats for a CFB team game-by-game for the season}
\usage{
.cfb_team_game_by_game(team_id, year, sprt = "MFB")
}
\arguments{
\item{team_id}{Team ID (form team_mapping)}

\item{year}{Select year, (example: 2015-2016 season is 2016)}

\item{sport}{Tradiational Sport Name}
}
\description{
Option to select data per season or per game
}
\examples{
.cfb_team_game_by_game(697,2018,sprt="MFB")

}
\keyword{Stats}
\keyword{Team}
