% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_base_schedule.R
\name{.extract_base_schedule}
\alias{.extract_base_schedule}
\title{Extract Raw Base schedule}
\usage{
.extract_base_schedule(team_id, year, sport)
}
\arguments{
\item{team_id}{Team ID (form team_mapping)}

\item{sprt}{Tradiational Sport Name}

\item{yr}{Select year, (example: 2015-2016 season is 2016)}
}
\description{
Extracts the date, team's played and the results
}
\examples{

}
\keyword{internal}
