% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player_game.R
\name{.player_game}
\alias{.player_game}
\title{Get Player's Gamy by Game Stats}
\usage{
.player_game(team_id, year, sport)
}
\arguments{
\item{team_id}{Team ID from team_mapping}

\item{year}{Actual Year
Note for seasons Year1-Year2, Year2 is what you want to use in the functions.
For example, the 2015-2016 season is denoted for 2016.}

\item{sport}{Tradiational Sport Name or Sport Code}
}
\description{
Returns a dataframe of stats for all players by game
in a specific sport and year.
}
\examples{
.player_game(697,2016,"MBB")

.player_game(697,2015,"Womens Soccer")

}
\keyword{internal}
