% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_weekly_team.R
\name{.raw_weekly_team}
\alias{.raw_weekly_team}
\title{Extract Raw Weekly Data
Includes both team and player data per game
This is further cleaned when the team or player functions are called}
\usage{
.raw_weekly_team(team_id, year, sport)
}
\arguments{
\item{team_id}{Team ID (form team_mapping)}

\item{sprt}{Tradiational Sport Name}

\item{yr}{Select year, (example: 2015-2016 season is 2016)}
}
\description{
Extracts raw game by game data.
}
\examples{

}
\keyword{internal}
