% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_game.R
\name{.team_game}
\alias{.team_game}
\title{Get Team Season Stats}
\usage{
.team_game(team_id, year, sport)
}
\arguments{
\item{team_id}{Team ID from team_mapping}

\item{year}{Actual Year
Note for seasons Year1-Year2, Year2 is what you want to use in the functions.
For example, the 2015-2016 season is denoted for 2016.}

\item{sport}{Tradiational Sport Name or Sport Code}
}
\description{
Returns a dataframe of the overall team's stats by game
for the sport and year chosen.
}
\examples{
team_game(697,2016,"MBB")

team_game(697,2015,"Womens Soccer")
}
\keyword{internal}
