% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_season_stats.R
\name{.team_season_stats}
\alias{.team_season_stats}
\title{Get Team Season Stats}
\usage{
.team_season_stats(team_id, year, sport)
}
\arguments{
\item{team_id}{Team ID from team_mapping}

\item{year}{Actual Year
Note for seasons Year1-Year2, Year2 is what you want to use in the functions.
For example, the 2015-2016 season is denoted for 2016.}

\item{sport}{Tradiational Sport Name or Sport Code}
}
\description{
Returns a dataframe of the overall team's season stats
for the sport and year chosen.
}
\examples{
team_season_stats(697,2016,"MBB")

team_season_stats(697,2015,"Womens Soccer")
}
\keyword{internal}
