% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_url_stats.R
\name{.team_url_stats}
\alias{.team_url_stats}
\title{Create Team URL's}
\usage{
.team_url_stats(team_id, year, sport)
}
\arguments{
\item{team_id}{Team ID from team_mapping}

\item{year}{Actual Year (example: 2016 or 2015)}

\item{sport}{Tradiational Sport Name or Sport Code}
}
\description{
Create's team urls for different names.
}
\examples{
team_url_stats(697,2015,"Men's Basketball")

team_url_stats(697,2015,"Womens Soccer")
}
\keyword{internal}
