% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/year_ref.R
\name{.year_ref}
\alias{.year_ref}
\title{Reference Year ID's}
\usage{
.year_ref(yr, sprt)
}
\arguments{
\item{yr}{Select year, (example: 2015-2016 season is 2016)}

\item{sprt}{Tradiational Sport Name}
}
\description{
The NCAA website uses different "year id's" for
each sport. This function cross-references based on sport and year selection.
}
\examples{
year_ref(2016,"mens basketball")
year_ref(2015,"MBB")

year_ref(2015,"MBB")
}
\keyword{internal}
