% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player_stats.R
\name{player_stats}
\alias{player_stats}
\title{Extract Player Stats for a team, year and sport.}
\usage{
player_stats(team_id, year, sport, by = "season")
}
\arguments{
\item{team_id}{Team ID (form team_mapping)}

\item{year}{Select year, (example: 2015-2016 season is 2016)}

\item{sport}{Tradiational Sport Name}

\item{by}{Stats over season, or stats per game}
}
\description{
Option to select data per season or per game
}
\examples{
player_stats(697,2016,"MBB",by="SEASON")

player_stats(697,2016,"WSO",by="game")


}
\keyword{Player}
\keyword{Stats}
