% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sport_code.R
\name{sport_code}
\alias{sport_code}
\title{Reference Sport ID's}
\usage{
sport_code(sprt)
}
\arguments{
\item{sprt}{Tradiational Sport Name}
}
\description{
The NCAA website uses different "sport id's" for
each sport. This function references those baesd on name.
}
\examples{
sport_code("mens basketball")
sport_code("MBB")

sport_code("Women's Soccer")
}
\keyword{Code}
\keyword{Lookup}
\keyword{Sport}
