% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sport_df}
\alias{sport_df}
\title{Sport Code Reference}
\format{A data frame with 17 rows and 2 variables:
\describe{
  \item{names}{Full name of sports}
  \item{id}{Sport code that NCAA uses to identify the Sport Names}
  ...
}}
\source{
\url{http://stats.ncaa.org}
}
\usage{
sport_df
}
\description{
A dataset mapping the sport names to the NCAA sport id's.
This is helpful in cross-refrencing what sport you want to get.
}
\keyword{datasets}
