% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_mapping.R
\name{team_mapping}
\alias{team_mapping}
\title{Team Name to Team ID Reference}
\usage{
team_mapping(yr, sprt)
}
\arguments{
\item{yr}{Actual Year (example: 2016 or 2015)}

\item{sprt}{Tradiational Sport Name or Sport Code}
}
\description{
The NCAA website has a unique identifier for each time in
each sport. This function references those baesd on name.
}
\examples{
team_mapping(2015,"Men's Basketball")
team_mapping(2015,"MBB")

team_mapping(2016,"Womens Soccer")
}
\keyword{Find}
\keyword{Teams}
