% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_schedule.R
\name{team_schedule}
\alias{team_schedule}
\title{Team Schedule and Results}
\usage{
team_schedule(team_id, year, sport)
}
\arguments{
\item{team_id}{Team ID from team_mapping}

\item{year}{Actual Year (example: 2016 or 2015)}

\item{sport}{Tradiational Sport Name or Sport Code}
}
\description{
Returns a dataframe of the team's schedule and results.
Additional information includes: (1) site and (2) cumulative record
}
\examples{
team_schedule(697,2016,"MBB")

team_schedule(697,2015,"Womens Soccer")

}
\keyword{Schedule}
\keyword{Team}
