% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_stats.R
\name{team_stats}
\alias{team_stats}
\title{Extract Team Stats for a year and sport.}
\usage{
team_stats(team_id, year, sport, by = "season")
}
\arguments{
\item{team_id}{Team ID (form team_mapping)}

\item{year}{Select year, (example: 2015-2016 season is 2016)}

\item{sport}{Tradiational Sport Name}

\item{by}{Stats over season, or stats per game}
}
\description{
Option to select data per season or per game
}
\examples{
team_stats(697,2016,"MBB",by="SEASON")

team_stats(697,2016,"WSO",by="game")

}
\keyword{Stats}
\keyword{Team}
