% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{year_df}
\alias{year_df}
\title{Season Mapping to Years}
\format{A data frame with 92 rows and 4 variables:
\describe{
  \item{season}{Sport Season}
  \item{id}{Year referencing the Season}
  \item{sport}{Sport Code}
  \item{year_id}{Year ID}
  ...
}}
\source{
\url{http://stats.ncaa.org}
}
\usage{
year_df
}
\description{
This dataframes helps identify the year_id for each sport and its season.
This year_id is used intermally to find the correct urls on NCAA stats website.
}
\keyword{datasets}
